added interfaces for map generator components
