+++ core/src/com/geometric/mapgenerators/MapGenerator.java
@@ -1,5 +1,8 @@
 package com.geometric.mapgenerators;
 
+/**
+ * generates random map with walls and ways between them
+ */
 public interface MapGenerator {
     GameMap generate(int mapWidth, int mapHeight);
 }
